Fix build performance problems on build
Running sitemap.register_resource_list_manipulator is quite costly on a sitemap
with many resources. So we need to make sure to call it as rarely as
possible. This change refactors our Image class into two classes. The
actual Image resource and the resource list manipulator (Manipulator)
that holds all images and can be called one time to add all image
resources.

This way building a large app is quite more performant.

Running middleman server with middleman-images enabled is still not that
performant, since we still have to run
register_resource_list_manipulator for all images on the page.